Move helpers to common file
