+++ src/types.ts
@@ -1,6 +1,11 @@
 import { OpenAPIV3 } from "openapi-types";
 import ts from "typescript";
-import { isReferenceObject, refToTypeName } from "./common";
+import {
+  appendNullToUnion,
+  isReferenceObject,
+  nonArraySchemaObjectTypeToTs,
+  refToTypeName,
+} from "./common";
 
 function isArraySchemaObject(
   param: OpenAPIV3.ReferenceObject | OpenAPIV3.SchemaObject
@@ -24,56 +29,6 @@ function schemaObjectTypeToArrayType(
   item: OpenAPIV3.NonArraySchemaObject
 ): ts.TypeNode {
   return appendNullToUnion(nonArraySchemaObjectTypeToTs(item), item.nullable);
-}
-
-function schemaObjectTypeToEnumType(enumValues: string[], nullable?: boolean) {
-  const enums = enumValues.map((value) =>
-    ts.factory.createLiteralTypeNode(ts.factory.createStringLiteral(value))
-  );
-
-  return appendNullToUnion(
-    ts.factory.createUnionTypeNode(/*types*/ enums),
-    nullable
-  );
-}
-
-function nonArraySchemaObjectTypeToTs(
-  item: OpenAPIV3.NonArraySchemaObject
-): ts.TypeNode {
-  if (!item.type) {
-    return ts.factory.createKeywordTypeNode(ts.SyntaxKind.AnyKeyword);
-  }
-
-  switch (item.type) {
-    case "string": {
-      if (item.enum) {
-        return schemaObjectTypeToEnumType(item.enum, item.nullable);
-      }
-      return appendNullToUnion(
-        ts.factory.createKeywordTypeNode(ts.SyntaxKind.StringKeyword),
-        item.nullable
-      );
-    }
-    case "integer":
-    case "number":
-      return appendNullToUnion(
-        ts.factory.createKeywordTypeNode(ts.SyntaxKind.NumberKeyword),
-        item.nullable
-      );
-    case "boolean":
-      return appendNullToUnion(
-        ts.factory.createKeywordTypeNode(ts.SyntaxKind.BooleanKeyword),
-        item.nullable
-      );
-    case "object": {
-      return appendNullToUnion(
-        createLiteralNodeFromProperties(item),
-        item.nullable
-      );
-    }
-    default:
-      return ts.factory.createKeywordTypeNode(ts.SyntaxKind.UnknownKeyword);
-  }
 }
 
 function resolveArray(item: OpenAPIV3.ArraySchemaObject): ts.TypeNode {
@@ -185,17 +140,6 @@ function createTypeAliasDeclarationTypeWithSchemaObject(
   }
 
   return nonArraySchemaObjectTypeToTs(item);
-}
-
-function appendNullToUnion(type: ts.TypeNode, nullable?: boolean) {
-  return nullable
-    ? ts.factory.createUnionTypeNode(
-        /*types*/ [
-          type,
-          ts.factory.createLiteralTypeNode(ts.factory.createNull()),
-        ]
-      )
-    : type;
 }
 
 export function createLiteralNodeFromProperties(item: OpenAPIV3.SchemaObject) {
